+++ pnlibrary-api/src/main/kotlin/ru/privatenull/pnlibrary/api/actions/SoundAction.kt
@@ -9,9 +9,9 @@ data class SoundAction(
     val source: Sound.Source = Sound.Source.MASTER,
     val volume: Float = 1f,
     val pitch: Float = 1f,
-    val target: Action.Target = Action.Target.PLAYER,
+    val target: ActionTarget = ActionTarget.PLAYER,
 ) : Action {
-    override fun execute(context: Action.Context) {
+    override fun execute(context: ActionContext) {
         val sound = Sound.sound(Key.key(key), source, volume, pitch)
         if (!context.target(target).playSound(sound)) {
             context.logger.warning("Sound action is unsupported by the current platform: $key")